by_script

+++ backend/getConfig.php
@@ -12,15 +12,19 @@
             }
             fclose($fp); 
             $json = json_decode($txt); 
-            if($order == 'read'){          
-                $out =$json->$field;
-                print json_encode($out);
-            }else{ 
-                if(!property_exists($json, $field)){
-                    $json->$field = new class{};
-                }               
-                $json->$field->$order = $value;                                        
-                return file_put_contents($path, json_encode($json));
+
+            if(property_exists($json, $field)){
+                if($order == 'read'){
+//                    echo $json->$field .'|';
+                    $out =$json->$field;
+                    print json_encode($out);
+                }else{ 
+                    if(!property_exists($json, $field)){
+                        $json->$field = new class{};
+                    }               
+                    $json->$field->$order = $value;                                        
+                    return file_put_contents($path, json_encode($json));
+                }    
             }
         }        
     }
